+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,41 @@
+---
+name: Bug report
+about: Report a reproducible bug.
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+<!-- Ignoring this template may result in your bug report getting deleted -->
+
+**Describe the bug**
+
+A clear and concise description of what the bug is. Before submitting, please remove unnecessary sections.
+
+**To Reproduce**
+
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. See error
+
+***Code***
+
+```js
+// code goes here
+```
+
+**Expected behavior**
+
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+
+If applicable, add screenshots to help explain your problem (drag and drop the image).
+
+**Platform:**
+
+ - Device: [Desktop, Mobile]
+ - OS: [Windows, MacOS, Linux, Android, iOS]
+ - Browser: [Chrome, Firefox, Safari, Edge]
