+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,5 @@
+blank_issues_enabled: true
+contact_links:
+  - name: Help and Support
+    url: https://discord.gg/w6CzHy35E2
+    about: Please ask the discord community if you have questions or need help.
