+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,25 @@
+---
+name: Feature request
+about: Suggest an idea for the project.
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+<!-- Ignoring this template may result in your feature request getting deleted -->
+
+**Is your feature request related to a problem or a need? Please describe.**
+
+A clear and concise description of what the problem is. Ex: 
+* I'm always frustrated when [...]
+* I would like to be able to [...]
+* I think [...] should be like [...] instead of [...]
+
+**Describe the solution you'd like**
+
+A clear and concise description of what you want to happen.
+
+**Additional context**
+
+Add any other context or screenshots about the feature request here.
