+++ docs/VirtualEnvironment_ClassDiagram.drawio
