+++ examples/base-template/index.html
@@ -12,7 +12,7 @@
     <script type="importmap"> { "imports": { "three": "https://cdn.skypack.dev/three@0.137.0" } } </script>
     <script async src="https://unpkg.com/es-module-shims@1.3.6/dist/es-module-shims.js"></script>
     <script type="module" src="./js/main.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/peerjs/1.4.5-rc.1/peerjs.min.js"></script>
+    <script async src="https://cdnjs.cloudflare.com/ajax/libs/peerjs/1.4.5-rc.1/peerjs.min.js"></script>
 </body>
 
 </html>
