+++ examples/classroom/assets/space-time.png
