+++ examples/classroom/authentication.html
