+++ examples/classroom/index.html
