+++ examples/classroom/js/authentication.js
