+++ examples/dome/index.html
