+++ examples/dome/js/main.js
@@ -1,0 +1,18 @@
+import { VirtualEnvironment } from '../../../js/VirtualEnvironment.js';
+
+// start by creating a basic virtual environment
+let virtualEnvironment = new VirtualEnvironment();
+
+// then fill your world with the stuff you want
+init();
+function init() {
+    virtualEnvironment.loadTerrain('../../../resources/terrains/dome.fbx', 0, -20, 0, "fbx", 0.2)
+    virtualEnvironment.spawnPlayer('../../../resources/avatars/yBot.glb', 0, 50, 0)
+}
+
+// then start the animation
+animate();
+function animate() {
+    virtualEnvironment.update();
+    requestAnimationFrame(animate);
+}
