+++ examples/floor-is-lava/js/main.js
@@ -8,32 +8,24 @@ let virtualEnvironment = new VirtualEnvironment();
 // then fill your world with the stuff you want
 init();
 function init() {
-    virtualEnvironment.spawnPlayer('../../../glb/avatars/yBot.glb', 0, 50, 0)
+    virtualEnvironment.spawnPlayer('../../../resources/avatars/yBot.glb', 0, 50, 0)
     
     setTimeout(() => {
-        virtualEnvironment.newSolidGeometriesFromSource('../../../glb/objects/rock_platform.glb', 0, 20, 1000, 16);
-        virtualEnvironment.newSolidGeometriesFromSource('../../../glb/objects/rock_platform.glb', 0, -20, -30, 16);
+        virtualEnvironment.newSolidGeometriesFromSource('../../../resources/objects/rock_platform.glb', 0, 20, 1000, 16);
+        virtualEnvironment.newSolidGeometriesFromSource('../../../resources/objects/rock_platform.glb', 0, -20, -30, 16);
 
         for(let i=0; i<35; i++) {
             let position = new THREE.Vector3(Math.random() * 800 - 400, 0, Math.random() * 1000);
-            virtualEnvironment.newSolidGeometriesFromSource('../../../glb/objects/rock_platform.glb', position.x, position.y, position.z, Math.random() * 20 + 4);
+            virtualEnvironment.newSolidGeometriesFromSource('../../../resources/objects/rock_platform.glb', position.x, position.y, position.z, Math.random() * 20 + 4);
         }
 
         virtualEnvironment.terrainController.generateCollider(virtualEnvironment.scene);
     }, 100);
 }
 
-if (!window.requestPostAnimationFrame) {
-    window.requestPostAnimationFrame = function(task) {
-        requestAnimationFrame(() => {
-            setTimeout(task, 0);
-        });
-    }
-}
-
+// then start the animation
+animate();
 function animate() {
-    requestPostAnimationFrame(animate);
-
     virtualEnvironment.update();
+    requestAnimationFrame(animate);
 }
-requestPostAnimationFrame(animate);
