+++ index.html
@@ -8,12 +8,12 @@
         <br>
         <p>here is a list of example implementations of the standard kit</p>
         <ol>
-            <li><a href="./exemples/classroom/authentication.html">Classroom</a></li>
-            <li><a href="./exemples/base-template/index.html">Base Template</a></li>
-            <li><a href="./exemples/spawn-planet/index.html">Spawn Planet</a></li>
-            <li><a href="./exemples/dome/index.html">Dome</a></li>
-            <li><a href="./exemples/floor-is-lava/index.html">Floor is Lava</a></li>
-            <li><a href="./exemples/terrain-generation/index.html">Terrain Generation</a></li>
+            <li><a href="./examples/classroom/authentication.html">Classroom</a></li>
+            <li><a href="./examples/base-template/index.html">Base Template</a></li>
+            <li><a href="./examples/spawn-planet/index.html">Spawn Planet</a></li>
+            <li><a href="./examples/dome/index.html">Dome</a></li>
+            <li><a href="./examples/floor-is-lava/index.html">Floor is Lava</a></li>
+            <li><a href="./examples/terrain-generation/index.html">Terrain Generation</a></li>
         </ol>
     </body>
 </html>
