+++ resources/objects/rock_platform.glb
