+++ resources/terrains/baseTemplate.fbx
